% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearingReceptor.R
\name{bearing}
\alias{bearing}
\title{Bearing of receptor from point source}
\usage{
bearing(sourceInput, receptorInput, sourceActivity = "primaryCrusher")
}
\arguments{
\item{sourceInput}{A data frame containing \code{sourceActivity} is a name for the pointsource, \code{type} (the type of material processing),
\code{x} for Easting, \code{y} for Northing and \code{z} for elevation height of the location.}

\item{receptorInput}{A data frame containing \code{receptor} the name/label for receptor point,
\code{x} for Easting, \code{y} for Northing and \code{z} for elevation height of the location.}

\item{sourceActivity}{The pointsource name such as \sQuote{primaryCrusher} or \sQuote{point A} which is
according to the \code{souceInput} \code{sourceActivity}.}
}
\value{
The value return to the direction of receptor point to the source location
}
\description{
\code{bearing} A function to determine the direction of receptor location to source location.
}
\details{
\code{bearing} require the source and receptor location in Easting and Northing format.
The \code{sourceActivity} for this model is the type of material processing activity in the quarry or
if the \code{sourceInput} is the quarries location, the \code{sourceActivity} is the name of the quarries.
}
\examples{
#demo
bearing(sourceInput = sourceInput, receptorInput = receptorInput, sourceActivity = "primaryCrusher")

}
\author{
Zul Fadhli & Dr Izhar Abadi
}
