% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blasting.R
\name{blasting}
\alias{blasting}
\title{Blasting}
\usage{
blasting(
  quarryInput,
  particleSize = "tsp",
  areaBlasted = quarryInput$areaBlasted,
  blastFreq = quarryInput$blastingFreq,
  blastHole.depth = quarryInput$blasthole,
  moistureContent = quarryInput$moisture
)
}
\arguments{
\item{quarryInput}{data frame containing column names: quarryProduction - quarry production tonne per hour,
blastingFreq - blasting frequency per month, areaBlasted - area blasted (km. sq), blasthole - average of depth of blast hole (m), drillhole - drilling hole per month,
moisture - percentage of moisture content, silt - percentage of silt content, unpavedDist - unpaved road distant (KM), pavedDist - paved road distanct (KM), 
unpavedVehicle - number of vehicle count travel using unpaved road per month, pavedVehicle - number of vehicle count travel using paved road per month, 
nPrimary - number of primary crusher, nSecondary - number of secondary crusher, nTertiary - number of tertiary crusher, nScreen - number of screen.}

\item{particleSize}{particle size of dust deposited rate. The default value is "tsp" for total suspended solid 
and other available particle size is "pm10" for particulate matter <10 micron.}

\item{areaBlasted}{a numeric value for area blasted (km. sq), or use the default to read areaBlasted value inside quarryInput dataframe.}

\item{blastFreq}{a numeric value for blasting frequency per month, or use the default to read blastFreq value inside quarryInput dataframe.}

\item{blastHole.depth}{a numeric value for average of depth of blast hole (m), or use the default to read blasthole value inside quarryInput dataframe.}

\item{moistureContent}{a numeric value for percentage of moisture content, or use the default to read moisture value inside quarryInput dataframe.}
}
\value{
A numeric value of estimation dust generation/\code{dustGenerated} for blasting activity.
}
\description{
A function to estimate dust generation from blasting activity.
}
\examples{
#demo
blasting(quarryInput, particleSize = "tsp")
}
\author{
Zul Fadhli & Dr Izhar Abadi
}
