% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmFreq.R
\name{calmFreq}
\alias{calmFreq}
\title{calmFreq}
\usage{
calmFreq(windInput, calm = TRUE)
}
\arguments{
\item{windInput}{A data frame containing \code{ws} for windspeed (m/s) and \code{wd} for wind direction}

\item{calm}{calm condition when wind speed below than 0.52 m/s. As default \code{calm} = \code{TRUE}. \code{FALSE} to exclude the calm condition.}
}
\value{
Total of calm condition in the dataset
}
\description{
Calm condition frequency count.
}
\examples{
#demo
calmFreq(windInput, calm =TRUE)
}
\author{
Zul Fadhli & Dr Izhar Abadi
}
