% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord2utm.R
\name{coord2utm}
\alias{coord2utm}
\title{coordiante to UTM format}
\usage{
coord2utm(location = sourceInput)
}
\arguments{
\item{location}{A data frame must contain column of "lat" for latitude and "long" for longitude. Value format in decimal.}
}
\description{
\code{coord2utm} Converting lat long into Northing and Easting. Using epsg 3375 GDM2000 (Malaysia) for map projection and datum transformation
}
\examples{
 
#demo
coord2utm(location = sourceInput)
}
\author{
Zul Fadhli & Dr Izhar Abadi
}
