% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pavedRoad.R
\name{pavedRoad}
\alias{pavedRoad}
\title{Paved Road dust generation}
\usage{
pavedRoad(
  quarryInput,
  particleSize = "tsp",
  vehicleTravelled = quarryInput$pavedVehicle,
  roadDistance = quarryInput$pavedDist,
  sp = 5,
  vehicleWeight = 10
)
}
\arguments{
\item{quarryInput}{data frame containing column names: quarryProduction - quarry production tonne per hour,
blastingFreq - blasting frequency per month, areaBlasted - area blasted (km. sq), blasthole - average of depth of blast hole (m), drillhole - drilling hole per month,
moisture - percentage of moisture content, silt - percentage of silt content, unpavedDist - unpaved road distant (KM), pavedDist - paved road distance (KM), 
unpavedVehicle - number of vehicle count travel using unpaved road per month, pavedVehicle - number of vehicle count travel using paved road per month, 
nPrimary - number of primary crusher, nSecondary - number of secondary crusher, nTertiary - number of tertiary crusher, nScreen - number of screen.}

\item{particleSize}{particle size of dust deposited rate. The default value is "tsp" for total suspended solid 
and other available particle size is "pm10" for particulate matter <10 micron.}

\item{vehicleTravelled}{number of vehicle used the road per month, use the default to read pavedVehicle value inside quarryInput dataframe}

\item{roadDistance}{total paved road in the quarry (KM), use the default to read pavedDist value inside quarryInput dataframe}

\item{sp}{a constant value. Default value is 5.}

\item{vehicleWeight}{average vehicle weight (tonne). The dafaults is 10.}
}
\description{
A function to estimate dust generated by paved road in quarry site.
}
\examples{
#demo
pavedRoad <- function(quarryInput, particleSize = "tsp", sp = 5, vehicleWeight = 10)
}
\author{
Zul Fadhli & Dr. Izhar Abadi
}
