% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graniteReceptorModel.R
\name{receptorModel.granite}
\alias{receptorModel.granite}
\title{Granite Receptor Model
multivariate equation for granite model
2.98 - 0.0123(AWS) - 0.00073(DSR) + 0.0006((ABA + TDH)/2) + 0.0089((NWB+WB)/2)}
\usage{
receptorModel.granite(
  sourceInput,
  receptorInput,
  windInput,
  sourceActivity = "primaryCrusher"
)
}
\arguments{
\item{sourceInput}{A data frame containing \code{sourceActivity} is a name for the pointsource, \code{type} (the type of material processing),
\code{x} for Easting, \code{y} for Northing and \code{z} for elevation height of the location.}

\item{receptorInput}{A data frame containing \code{receptor} the name/label for receptor point,
\code{x} for Easting, \code{y} for Northing and \code{z} for elevation height of the location.}

\item{windInput}{A data frame containing \code{ws} for windspeed (m/s) and \code{wd} for wind direction}

\item{sourceActivity}{The pointsource name such as \sQuote{primaryCrusher} or \sQuote{point A} which is
according to the \code{souceInput} \code{sourceActivity}.}
}
\value{
The results will be the estimation of dust deposited rate at all the receptors from a pointsource.
The unit measurement for dust deposited rate is ug/m2/month.
}
\description{
\code{receptorModel.granite} Multimaviate model established to estimate dust deposited rate at receptor
}
\details{
\code{receptorModel.granite} require the source and receptor location in Easting and Northing format.
The \code{sourceActivity} for this model is the type of material processing activity in the quarry or
if the \code{sourceInput} is the quarries location, the \code{sourceActivity} is the name of the quarries.
}
\examples{

#demo
receptorModel.granite(sourceInput, receptorInput, windInput, sourceActivity = "primaryCrusher")

}
\author{
Zul Fadhli & Izhar Abadi
}
