% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondaryCrusher.R
\name{secondaryCrusher}
\alias{secondaryCrusher}
\title{Dust Emission from Secondary crusher}
\usage{
secondaryCrusher(
  quarryInput,
  particleSize = "tsp",
  QuarryProduction = quarryInput$quarryProduction,
  crusherCount = quarryInput$nSecondary
)
}
\arguments{
\item{quarryInput}{data frame containing column names: quarryProduction - quarry production tonne per hour,
blastingFreq - blasting frequency per month, areaBlasted - area blasted (km. sq), blasthole - average of depth of blast hole (m), drillhole - drilling hole per month,
moisture - percentage of moisture content, silt - percentage of silt content, unpavedDist - unpaved road distant (KM), pavedDist - unpaved road distance (KM), 
unpavedVehicle - number of vehicle count travel using unpaved road per month, pavedVehicle - number of vehicle count travel using unpaved road per month, 
nPrimary - number of primary crusher, nSecondary - number of secondary crusher, nTertiary - number of tertiary crusher, nScreen - number of screen.}

\item{particleSize}{particle size of dust deposited rate. The default value is "tsp" for total suspended solid 
and other available particle size is "pm10" for particulate matter <10 micron.}

\item{QuarryProduction}{Numeric value for quarry production tonne per hour. Use default to read the value in quarryInput dataframe.}

\item{crusherCount}{Numeric value for the active Secondary crusher unit in the quarry location. Use default to read the value in quarryInput dataframe.}
}
\description{
A function to estimate dust generation from Secondary crusher activity.
}
\examples{
#demo 
secondaryCrusher(quarryInput, particleSize = "tsp", QuarryProduction = quarryInput$quarryProduction, crusherCount = quarryInput$nSecondary)
}
\author{
Zul Fadhli & Dr. Izhar Abadi
}
