% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us.R
\name{us}
\alias{us}
\title{Ground level wind speed at point source}
\usage{
us(
  sourceInput,
  windInput,
  sourceActivity = "primaryCrusher",
  AnemometerHeight = 2,
  stability = "A",
  windClass = "class1"
)
}
\arguments{
\item{sourceInput}{A data frame containing \code{sourceActivity} is a name for the pointsource, \code{type} (the type of material processing),
\code{x} for Easting, \code{y} for Northing and \code{z} for elevation height of the location.}

\item{windInput}{A data frame containing \code{ws} for windspeed (m/s) and 
\code{wd} for wind direction}

\item{sourceActivity}{Select the point source labeled in the \code{sourceInput} dataframe. For example; 
The pointsource name such as \sQuote{primaryCrusher} or \sQuote{point A} which is according to the \code{souceInput} \code{sourceActivity}.}

\item{AnemometerHeight}{Height of anemometer installation from sea level.}

\item{stability}{Six category stability which is 'A', 'B', 'C', 'D', 'E', and 'F'.}

\item{windClass}{Five wind band which is > 5.67 m/s, > 3.61 m/s, > 2.07 m/s, 0.52 and < 0.51
characterized as 'class1', 'class2', 'class3', 'class4' and 'calm' respectively}
}
\description{
A function to estimate wind speed at ground level of selected point source
}
\examples{
#demo
us <- function(sourceInput, windInput, sourceActivity = "primaryCrusher", AnemometerHeight = 2, stability = "A", windClass = "class1")
}
\author{
Zul Fadhli & Dr. Izhar Abadi
}
