% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windAvg.R
\name{windAvg}
\alias{windAvg}
\title{Mean of Wind speed by wind stability group}
\usage{
windAvg(windInput, stability = "F")
}
\arguments{
\item{windInput}{A data frame containing \code{ws} for windspeed (m/s) and 
\code{wd} for wind direction}

\item{stability}{Six category stability which is 'A', 'B', 'C', 'D', 'E', and 'F'.}
}
\description{
An utility function for MyQDMi apps
}
\examples{
 
#demo
windAvg(windInput, stability = "F")
}
\author{
Zul Fadhli & Dr. Izhar Abadi
}
