% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windFreq.R
\name{windFreq}
\alias{windFreq}
\title{Wind frequency by wind stability group, wind class, and wind direction}
\usage{
windFreq(
  windInput,
  stability = "F",
  windClass = "class1",
  windDirection = "NE"
)
}
\arguments{
\item{windInput}{}

\item{stability}{Six category stability which is 'A', 'B', 'C', 'D', 'E', and 'F'.}

\item{windClass}{Five wind band which is > 5.67 m/s, > 3.61 m/s, > 2.07 m/s, 0.52 and < 0.51
characterized as 'class1', 'class2', 'class3', 'class4' and 'calm' respectively}

\item{windDirection}{Wind direction category 'N', 'NE', 'E', 'SE', 'SW', 'W' and 'NW'}
}
\description{
An utility function for MyQDMi apps
}
\examples{
#demo
windFreq(windInput, stability = "F", windClass = "class1", windDirection = "NE")
}
\author{
Zul fadhli & Dr. Izhar Abadi
}
