% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance}
\usage{
distance(sourceInput, receptorInput, sourceActivity)
}
\arguments{
\item{sourceInput}{A data frame containing \code{sourceActivity} is a name for the pointsource, \code{type} (the type of material processing),
\code{x} for Easting, \code{y} for Northing and \code{z} for elevation height of the location.}

\item{receptorInput}{A data frame containing \code{receptor} the name/label for receptor point,
\code{x} for Easting, \code{y} for Northing and \code{z} for elevation height of the location.}

\item{sourceActivity}{}
}
\value{
distance between receptor location to source location
}
\description{
Distance
}
\examples{
#demo
distance(sourceInput, receptorInput, sourceActivity)
}
